app/requirements.txt
requests==2.31.0
python-dotenv==1.0.1
prometheus-client==0.20.0
flask==3.0.3